Move `"gui/app/layout"` into `"components/App.vue"`
- delete `src/app/gui/app/layout.js`
- move global `$.LayoutManager` definition within `components/App.vue`
